requirements.txt
# UI
streamlit==1.37.0
python-dotenv==1.0.1

# LangChain stack
langchain==0.2.16
langchain-core==0.2.38
langchain-community==0.2.15
langchain-text-splitters==0.2.4
langchain-groq==0.1.9

# Vector DB + Embeddings
chromadb==0.5.5
sentence-transformers==2.7.0
torch==2.4.0

# Document ingestion
pypdf==4.2.0
unstructured==0.15.7
python-docx==1.1.2
